Refactored the preferences stuff so pretty much anything to do with preferences lives in the preference package.
Added a preference manager to the preference package so you can now retrieve the value of any preference more easily - i.e by getting it from the preference manager rather than mucking around with the preference store directly.

Added homsite and dreamweaver compatibility preferences to the preference page.

Added single line block indent and implemented it as a dreamweaver and homesite preference. If either Homesite or Dreamweaver compatibility is enabled it does a block indent if part of a line is selected. Otherwise it replaces the selection with the current indent string.