I found the scopes.xml definition to be limiting, so I created a new tag called <types>, with a nested sub-elements of <type>.
It's just a more structured way of describing the scopes variables, because I wanted it to actually group the scopes as they really are in structures or queries, and give them a few attributes to describe special qualities.

So for example something like the "server" scope, is really just a specialized struct, with sub structs for "coldfusion" and "server".

"CFDirectory" scope is really just a specialized query, so it base type is "query"

"CGI" is a basic struct too, but it is immutable, nothing can be changed in it. So it has the attribute mutable="false".

Another attribute describes how the scope comes into being, there are 4
possible values:

- creation="global": available globally anywhere (application, server,
session, client, cookie, cgi, url, form, request)
- creation="auto": created automatically on demand (thistag, cfmx6:
cffile, cfhttp, cfstoredproc)
- creation="attribute": returned from an attribute name when calling a tag
(cfquery, cfdirectory, cfsearch, cfcollection, cfregistry, etc)
- creation="autoOrAttribute": returned from an attribute name, or if
omitted, returned automatically, which is an important distinction for
cfmx7 since now these scopes can be renamed (cffile, cfstoredproc,
cfftp, cfhttp)

This types.xml file also represents a lot of the extra types I added to the
list of cfbasetypes in dictionary2.xsd, for values like request, session, cgi, cfdirectory, cffile, and many of the others.

I think by having these scopes/types defined in this way will allow variable insight dialogs in the editor because they can then be looked as a valid coldfusion type, linked from a specific tag like <cffile> with the new <return> tag. When the user has one of these structures or queries variables returned from a tag, they could then use a variable insight to choose from the available fields to display or use.