fixes #161
from Paul Vernon:
Here's another little patch which addresses an issue with incorrectly
highlighted <cfhttp commands where url is present but it states it isn't and
is compulsory.

There are two things in this patch, firstly, it fixes the cfhttp url
parameter issue, secondly, it addresses a kind of typo in the error messages
by removing the hard coded "cf" string.

The cfhttp url parameter issue was weird to say the least. When tracing
through, if I had the code in uppercase it failed because it was looking for
lowercase "url" and if it was in lowercase, it was looking for uppercase
"URL"... In the end I just made it look for either which will negate the
issue altogether regardless of how the code is written.