added additional folding tags
added code for additional folding tags...there is now 39 explicitly stated with the ability for the user to change them and change the settings.
also adjusted code that has CFScript listed twice in the original 8.
Ideally, this would loop through each, instead of having each one defined like this.
Also, ideally, this would write/read to/from an external ini file, possibly in XML that a user could hack at if they want and that would be what is looped over instead of a lit of variables.  
Just a general thought, but this is a step and something I personally would find very beneficial.
Please publish to Eclipse Marketplace.